Make tor and phs demos more robust. Do not copy c.xxx.1 to c.xxx if c.xxx already exists with copydemo().
git-svn-id: https://svn.code.sf.net/p/auto-07p/code/trunk/auto/07p@2489 ee9e63dd-6847-0410-88cd-b7174c777ff7

+++ python/AUTOCommands.py
@@ -122,7 +122,8 @@ def copydemo(name):
             shutil.copy(f, ".")
         except IOError:
             pass
-    if os.path.exists("c.%s.1"%name):
+    if (os.path.exists(os.path.join(demodir,"c.%s.1"%name)) and
+        not os.path.exists(os.path.join(demodir,"c.%s"%name))):
         shutil.copy("c.%s.1"%name,"c.%s"%name)
     info("Copying demo %s ... done\n"%name)
 commandCopyDemo = command(copydemo,SIMPLE,"demo")
